upd: contacts.json moved to db folder
