+++ controllers/contacts-operations/refreshContacts.js
@@ -1,5 +1,5 @@
 const fs = require("fs/promises");
-const { contactsPath } = require("../../model");
+const { contactsPath } = require("../../db");
 
 const refreshContacts = async (contactsToFile) => {
   const newContacts = await fs.writeFile(
